CSS Updated
Updated to new css Styling and fixed some page organization.
