+++ src/App.css
@@ -1,5 +1,11 @@
 *{overflow: hidden;}
 
+body{
+    background: #282c34;
+    font-family:'Roboto',sans-serif;
+
+}
+
 a {
     color: inherit;
     text-decoration: none;
@@ -26,21 +32,3 @@ a:hover {
     }
 }
 
-.Landing {
-    color: white;
-    background: linear-gradient(-70deg, rgb(110,94,254) 30%, rgb(0,0,0) 30%);
-    display: flex;
-    justify-content:  center;
-    flex-direction: column;
-    align-items: center;
-    min-height: 85vh;
-    font-size: calc(10px + 2vmin);
-}
-
-.Landing-header {
-    flex-basis: 100%;
-}
-
-.Landing-p {
-    width: 70vw;
-}
