+++ src/components/Footer/Footer.css
@@ -1,9 +1,9 @@
 .FooterItems {
-    background: linear-gradient(90deg, rgb(110,94,254) 0%, rgba(73,63,252,1) 100%);
+    background: linear-gradient(90deg, rgb(157,101,201) 0%, rgb(122, 74, 158) 100%);
     position: fixed;
     bottom: 0;
     width: 100%;
-    height: 10vh;
+    height: 9vh;
     display: flex;
     flex-wrap: wrap;
     justify-content:  center;
