+++ src/components/LoginButton/LoginButton.css
@@ -4,16 +4,15 @@
     color: white;
     text-decoration: none;
     padding: 0.5rem 1rem;
-    background-color: #6d76f7;
+    background-color: #282c34;
     box-shadow: 0 1.75rem 3.5rem rgba(0, 0, 0, 0.1);
-    border-radius: 20px;
+    border:0.2em solid #FFFFFF;
+    font-family:'Roboto',sans-serif;
     width: 14ch;
 }
 
 .login-button:hover {
-    background-color: #6d76f7;
-    text-decoration: none;
-    padding: 0.5rem 1rem;
-    border-radius: 4px;
+    color:#000000;
+    background-color:#FFFFFF;
     transition: all 0.2s ease-out;
 }
