+++ src/components/Navbar/Navbar.css
@@ -1,5 +1,5 @@
 .NavbarItems {
-    background: linear-gradient(90deg, rgb(110,94,254) 0%, rgba(73,63,252,1) 100%);
+    background: linear-gradient(90deg, rgb(157,101,201) 0%, rgb(122, 74, 158) 100%);
     height: 6vh;
     display: flex;
     justify-content:  center;
@@ -10,7 +10,7 @@
 .navbar-logo {
     color: #fff;
     justify-self: start;
-    margin-left: 20px;
+    margin-left: 25px;
     cursor: pointer;
 }
 
@@ -34,12 +34,19 @@
     color: white;
     text-decoration: none;
     padding: 0.5rem 1rem;
+    font-family:'Roboto',sans-serif;
+    height: 100px;
 }
 
 .nav-links:hover {
-    background-color: #6d76f7;
-    border-radius: 4px;
-    transition: all 0.2s ease-out;
+    
+    
+    
+    padding: 20px;
+    width: 200px;
+    height: 150px;
+    
+    transition: all 0.2s ease-in-out;
 }
 
 .signup-link{
