+++ src/pages/About/About.css
@@ -1,0 +1,28 @@
+.main {
+    
+    display: flex;
+    justify-content:  center;
+    flex-direction: column;
+    align-items: center;
+    min-height: 85vh;
+    font-size: calc(10px + 2vmin);
+ }
+
+ .header {
+    color: white;
+    flex-basis: 100%;
+    font-family:'Roboto',sans-serif;
+    text-align: center;
+    vertical-align: middle;
+    font-weight: 600;
+    font-size: 80px;
+    width: 30vw;
+    
+}
+.p {
+    color: white;
+    width: 50vw;
+    font-family:'Roboto',sans-serif;
+    font-size: 25px;
+    
+}
