+++ src/pages/Contributors/Contributors.css
@@ -1,15 +1,44 @@
 .ContributorItems {
-	min-height: 85vh;
-	background: url(cont-background.png) center no-repeat;
+	display: flex;
+	background: #282c34;
+	justify-content: center;
+	flex-direction: column;
+	align-items: center;
+	min-height: 75vh;
+}
+
+.ContributorItems h1{
+	color: white;
+    
+    font-size: 55px;
+}
+.ContributorItems h2{
+	color: white;
+    font-size: 30px;
+}
+.ContributorItems p{
+	color: white;
+    width: 25vw;
+    font-size: 25px;
 }
 
 .Contributor-Link {
-	color: silver;
-	text-decoration: none;
+	font-size: x-large;
+    display:inline-block;
+    padding:0.35em .5em;
+    border:0.1em solid #FFFFFF;
+    margin:0 0.3em 0.3em 0;
+    border-radius:0.12em;
+    box-sizing: border-box;
+    text-decoration:none;
+    font-family:'Roboto',sans-serif;
+    font: weight 10px;
+    color:#FFFFFF;
+    text-align:center;
+    transition: all 0.2s;
 }
 
 .Contributor-Link:hover {
-	color: black;
-	text-decoration: none;
-	cursor: pointer;
+    color:#000000;
+    background-color:#FFFFFF;
 }
