+++ src/pages/Error404/Error404.css
@@ -1,13 +1,6 @@
 *{overflow: hidden;}
 
-body{
-    margin: 0;
-    padding: 0;
-    min-height: 100vh;
-    font-family: "montserrat", sans-serif;
-    background-image: linear-gradient(125deg, #6a89cc, #b8e994);
-    
-}
+
 
 .Error{
 
@@ -17,10 +10,15 @@ body{
     transform: translateY(-50%);
     text-align: center;
     color: #343434;
+    margin: 0;
+    padding: 0;
+    min-height: 100vh;
+    font-family: "montserrat", sans-serif;
+    background-image: linear-gradient(125deg, #9d65c9, #282c34);
 }
 
 .Error h1{
-    font-size: 200px;;
+    font-size: 200px;
     margin: 0;
     font-weight: 900;
     letter-spacing: 20px;
@@ -29,7 +27,7 @@ body{
     -webkit-background-clip: text;
 }
 .Error h2{
-    font-size: 50px;;
+    font-size: 50px;
     margin: 0;
     font-weight: 300;
     letter-spacing: 1px;
